+++ packages/atomic/cypress/integration/result-list/result-components/result-number-selectors.ts
@@ -3,6 +3,7 @@ import {ResultListSelectors} from '../result-list-selectors';
 export const resultNumberComponent = 'atomic-result-number';
 
 export const ResultNumberSelectors = {
+  shadow: () => cy.get(resultNumberComponent),
   firstInResult: () =>
     ResultListSelectors.firstResult().find(resultNumberComponent),
   formats: {