+++ packages/quantic/cypress/integration/results-per-page/results-per-page-actions.ts
@@ -1,0 +1,7 @@
+import {ResultsPerPageSelectors} from './results-per-page-selectors';
+
+export const ResultsPerPageActions = {
+  selectValue: (value: number) => {
+    ResultsPerPageSelectors.choice().contains(value).click();
+  },
+};