+++ packages/quantic/force-app/main/default/lwc/quanticNumericFacet/quanticNumericFacet.js
@@ -2,6 +2,7 @@ import {LightningElement, track, api} from 'lwc';
 import {
   registerComponentForInit,
   initializeWithHeadless,
+  getHeadlessBindings
 } from 'c/quanticHeadlessLoader';
 import {I18nUtils} from 'c/quanticUtils';
 import LOCALE from '@salesforce/i18n/locale';
@@ -10,11 +11,22 @@ import clearFilter from '@salesforce/label/c.quantic_ClearFilter';
 import clearFilter_plural from '@salesforce/label/c.quantic_ClearFilter_plural';
 import collapseFacet from '@salesforce/label/c.quantic_CollapseFacet';
 import expandFacet from '@salesforce/label/c.quantic_ExpandFacet';
+import min from "@salesforce/label/c.quantic_Min";
+import max from "@salesforce/label/c.quantic_Max";
+import numberInputMinimum from "@salesforce/label/c.quantic_NumberInputMinimum";
+import numberInputMaximum from "@salesforce/label/c.quantic_NumberInputMaximum";
+import apply from "@salesforce/label/c.quantic_Apply";
+import numberInputApply from "@salesforce/label/c.quantic_NumberInputApply";
+import messageWhenRangeOverflow from '@salesforce/label/c.quantic_MessageWhenRangeOverflow';
+import messageWhenRangeUnderflow from '@salesforce/label/c.quantic_MessageWhenRangeUnderflow';
 
 /** @typedef {import("coveo").NumericFacetState} NumericFacetState */
+/** @typedef {import("coveo").NumericFilterState} NumericFilterState*/
 /** @typedef {import("coveo").NumericFacet} NumericFacet */
-/** @typedef {import("coveo").SearchEngine} SearchEngine */
+/** @typedef {import("coveo").NumericFilter} NumericFilter */
 /** @typedef {import("coveo").NumericFacetValue} NumericFacetValue */
+/** @typedef {import("coveo").SearchStatus} SearchStatus */
+/** @typedef {import("coveo").SearchEngine} SearchEngine */
 
 /**
  * The `QuanticNumericFacet` component displays facet values as numeric ranges.
@@ -86,6 +98,15 @@ export default class QuanticNumericFacet extends LightningElement {
     item.end
   )}`;
   /**
+   * Whether this facet should contain an input allowing users to set custom ranges.
+   * Depending on the field, the input can allow either decimal or integer values.
+   *   - `integer`
+   *   - `decimal`
+   * @api
+   * @type {'integer' | 'decimal'}
+   */
+  @api withInput;
+  /*
    * Whether the facet is collapsed.
    * @api
    * @type {boolean}
@@ -102,17 +123,50 @@ export default class QuanticNumericFacet extends LightningElement {
 
   /** @type {NumericFacetState} */
   @track state;
+  /** @type {NumericFilterState} */
+  @track filterState = {
+    isLoading: false,
+    facetId: undefined
+  }
 
   /** @type {NumericFacet} */
   facet;
+  /**  @type {NumericFilter} */
+  numericFilter;
+  /**  @type {SearchStatus} */
+  searchStatus;
+  /** @type {boolean} */
+  showPlaceholder = true;
   /** @type {Function} */
   unsubscribe;
+  /** @type {Function} */
+  unsubscribeFilter;
+  /** @type {Function} */
+  unsubscribeSearchStatus;
+  /** @type {boolean} */
+  usingManualRange=false;
+
+  /** @type {string} */
+  start;
+  /** @type {string} */
+  end;
+
+  minSafeValue = Number.MIN_VALUE;
+  maxSafeValue = Number.MAX_VALUE;
 
   labels = {
     clearFilter,
     clearFilter_plural,
     collapseFacet,
     expandFacet,
+    min,
+    max,
+    numberInputMinimum,
+    numberInputMaximum,
+    apply,
+    numberInputApply,
+    messageWhenRangeOverflow,
+    messageWhenRangeUnderflow
   };
 
   connectedCallback() {
@@ -127,6 +181,11 @@ export default class QuanticNumericFacet extends LightningElement {
    * @param {SearchEngine} engine
    */
   initialize = (engine) => {
+    this.searchStatus = CoveoHeadless.buildSearchStatus(engine);
+    this.unsubscribeSearchStatus = this.searchStatus.subscribe(() =>
+      this.updateState()
+    );
+
     this.facet = CoveoHeadless.buildNumericFacet(engine, {
       options: {
         field: this.field,
@@ -137,15 +196,39 @@ export default class QuanticNumericFacet extends LightningElement {
         facetId: this.facetId ?? this.field,
       }
     });
+    if(this.withInput) {
+      this.initializeFilter(engine);
+    }
+    this.searchStatus = CoveoHeadless.buildSearchStatus(engine);
     this.unsubscribe = this.facet.subscribe(() => this.updateState());
+    this.unsubscribeSearchStatus = this.searchStatus.subscribe(() => this.updateState());
+  }
+
+  /**
+   * @param {import("coveo").SearchEngine} engine
+   */
+  initializeFilter(engine) {
+     this.numericFilter = CoveoHeadless.buildNumericFilter(engine, {
+      options: {
+        field: this.field,
+        facetId: this.facet.state.facetId ? `${this.facet.state.facetId}_input` : undefined
+      }
+    });
+    this.unsubscribeFilter = this.numericFilter.subscribe(() => this.updateState());
   }
 
   disconnectedCallback() {
     this.unsubscribe?.();
+    this.unsubscribeFilter?.();
+    this.unsubscribeSearchStatus?.();
   }
 
   updateState() {
-    this.state = this.facet.state;
+    this.state = this.facet?.state;
+    this.showPlaceholder = !this.searchStatus?.state?.hasError && !this.searchStatus?.state?.firstSearchExecuted;
+    this.filterState = this.numericFilter?.state;
+    this.start = this.filterState?.range?.start?.toString();
+    this.end = this.filterState?.range?.end?.toString();
   }
 
   get values() {
@@ -161,12 +244,25 @@ export default class QuanticNumericFacet extends LightningElement {
     );
   }
 
+  get step() {
+    return this.withInput ==='integer' ? '1' : 'any';
+  }
+  /** @returns {HTMLInputElement} */
+  get inputMin() {
+    return this.template.querySelector('.numeric__input-min');
+  }
+
+  /** @returns {HTMLInputElement} */
+  get inputMax() {
+    return this.template.querySelector('.numeric__input-max');
+  }
+
   get hasValues() {
     return this.values.length !== 0;
   }
 
   get hasActiveValues() {
-    return this.state?.hasActiveValues;
+    return this.state?.hasActiveValues || this.filterState?.range;
   }
 
   get actionButtonIcon() {
@@ -183,7 +279,11 @@ export default class QuanticNumericFacet extends LightningElement {
   }
 
   get numberOfSelectedValues() {
-    return this.state.values.filter(({state}) => state === 'selected').length;
+    return this.filterState?.range ? 1 : this.state.values.filter(({state}) => state === 'selected').length;
+  }
+
+  get showValues() {
+    return !this.searchStatus?.state?.hasError && (!this.usingManualRange || !this.filterState?.range) && !!this.values.length;
   }
 
   get clearFilterLabel() {
@@ -194,15 +294,52 @@ export default class QuanticNumericFacet extends LightningElement {
     return '';
   }
 
-  /**
-   * @param {CustomEvent<NumericFacetValue>} evt
+  onChangeMin(evt) {
+    evt.preventDefault();
+    this.start = evt.target.value;
+  }
+
+  onChangeMax(evt) {
+    evt.preventDefault();
+    this.end = evt.target.value;
+  }
+
+  setValidityParameters() {
+    this.inputMin.max = this.end || this.maxSafeValue.toString();
+    this.inputMax.min = this.start || this.minSafeValue.toString();
+    this.inputMin.required = true;
+    this.inputMax.required = true;
+  }
+
+  resetValidityParameters() {
+    this.inputMin.max = this.maxSafeValue.toString();
+    this.inputMax.min = this.minSafeValue.toString();
+    this.inputMin.required = false;
+    this.inputMax.required = false;
+  }
+
+  /** 
+   * @param {CustomEvent<{value: string}>} evt
    */
-  onSelect(evt) {
-    this.facet.toggleSelect(evt.detail);
+  onSelectValue(evt) {
+    const item = this.values.find((value) => this.formattingFunction(value) === evt.detail.value);
+    this.facet.toggleSelect(item);
+    this.usingManualRange = false;
   }
 
   clearSelections() {
-    this.facet.deselectAll();
+    this.usingManualRange = false;
+    if(this.filterState?.range) {
+      this.numericFilter.clear();
+    }
+    this.resetValidityParameters();
+    this.facet?.deselectAll();
+    [...this.template.querySelectorAll('lightning-input')].forEach( input => {
+      // @ts-ignore
+      input.checkValidity();
+      // @ts-ignore
+      input.reportValidity();
+    });
   }
 
   toggleFacetVisibility() {
@@ -211,5 +348,50 @@ export default class QuanticNumericFacet extends LightningElement {
 
   preventDefault(evt) {
     evt.preventDefault();
+  }
+
+  onApply(evt) {
+    evt.preventDefault();
+
+    this.setValidityParameters();
+
+    const allValid = [...this.template.querySelectorAll('lightning-input')]
+      .reduce((validSoFar, inputCmp) => {
+        // @ts-ignore
+        return validSoFar && inputCmp.reportValidity();
+      }, true);
+      
+    this.resetValidityParameters();
+
+    if (!allValid) {
+      return;
+    }
+    this.usingManualRange = true;
+    const engine = getHeadlessBindings(this.engineId).engine;
+    engine.dispatch(CoveoHeadless.loadNumericFacetSetActions(engine).deselectAllNumericFacetValues(this.facet.state.facetId));
+    this.numericFilter.setRange({
+      start: this.inputMin ? Number(this.inputMin.value) : undefined,
+      end: this.inputMax ? Number(this.inputMax.value) : undefined
+    });
+  }
+
+  get numberInputMinimumLabel() {
+    return I18nUtils.format(this.labels.numberInputMinimum, this.label);
+  }
+
+  get numberInputMaximumLabel() {
+    return I18nUtils.format(this.labels.numberInputMaximum, this.label);
+  }
+
+  get numberInputApplyLabel() {
+    return I18nUtils.format(this.labels.numberInputApply, this.label);
+  }
+
+  get customMessageOverflow() {
+    return I18nUtils.format(this.labels.messageWhenRangeOverflow, this.end);
+  }
+  
+  get customMessageUnderflow() {
+    return I18nUtils.format(this.labels.messageWhenRangeUnderflow, this.start);
   }
 }