+++ packages/quantic/force-app/main/default/lwc/quanticResultList/quanticResultList.js
@@ -5,6 +5,7 @@ import {registerComponentForInit, initializeWithHeadless} from 'c/quanticHeadles
 /** @typedef {import("coveo").ResultList} ResultList */
 /** @typedef {import("coveo").ResultListState} ResultListState */
 /** @typedef {import("coveo").ResultTemplatesManager} ResultTemplatesManager */
+/** @typedef {import("coveo").SearchStatus} SearchStatus */
 /** @typedef {import("coveo").SearchEngine} SearchEngine */
 
 /**
@@ -24,17 +25,25 @@ export default class QuanticResultList extends LightningElement {
    * A list of fields to include in the query results, separated by commas.
    * @api
    * @type {string}
-   * @defaultValue `'date,author,source,language,filetype,parents'`
+   * @defaultValue `'date,author,source,language,filetype,parents,sfknowledgearticleid'`
    */
-  @api fieldsToInclude = 'date,author,source,language,filetype,parents';
+  @api fieldsToInclude = 'date,author,source,language,filetype,parents,sfknowledgearticleid';
 
   /** @type {ResultListState}*/
   @track state;
 
   /** @type {ResultList} */
   resultList;
+  /** @type {boolean} */
+  showPlaceholder = true;
+  /** @type {number} */
+  numberOfResults = 10;
   /** @type {Function} */
   unsubscribe;
+  /** @type {Function} */
+  unsubscribeSearchStatus;
+  /** @type {Function} */
+  unsubscribeResultsPerPage;
   /** @type {ResultTemplatesManager} */
   resultTemplatesManager;
 
@@ -50,6 +59,14 @@ export default class QuanticResultList extends LightningElement {
    * @param {SearchEngine} engine
    */
   initialize = (engine) => {
+    this.resultsPerPage = CoveoHeadless.buildResultsPerPage(engine);
+    this.unsubscribeResultsPerPage = this.resultsPerPage.subscribe(() => this.updateState());
+
+    this.searchStatus = CoveoHeadless.buildSearchStatus(engine);
+    this.unsubscribeSearchStatus = this.searchStatus.subscribe(() =>
+      this.updateState()
+    );
+
     this.resultList = CoveoHeadless.buildResultList(engine, {
       options: {
         fieldsToInclude: this.fields
@@ -73,15 +90,23 @@ export default class QuanticResultList extends LightningElement {
 
   disconnectedCallback() {
     this.unsubscribe?.();
+    this.unsubscribeSearchStatus?.();
+    this.unsubscribeResultsPerPage?.();
   }
 
   updateState() {
-    this.state = this.resultList.state;
+    this.state = this.resultList?.state;
+    this.numberOfResults = this.resultsPerPage?.state?.numberOfResults;
+    this.showPlaceholder = !this.searchStatus?.state?.hasError && !this.searchStatus?.state?.firstSearchExecuted && !!this.numberOfResults;
   }
 
   get fields() {
     if (this.fieldsToInclude.trim() === '') return [];
     return this.fieldsToInclude.split(',').map((field) => field.trim());
+  }
+
+  get hasResults() {
+    return !!this.results.length;
   }
 
   get results() {